dbt_airflow_snowflake/requirements.txt
dbt
dbt-snowflake
astronomer-cosmos
apache-airflow-providers-snowflake
apache-airflow[amazon]
pandas
boto3

